rewrite providers structures
